+++ lib/features/search/search_screen.dart
@@ -200,7 +200,7 @@ class _SearchScreenState extends ConsumerState<SearchScreen> {
                           fit: StackFit.expand,
                           children: [
                             Image.file(
-                              File(photo.filePath),
+                              File(photo.resolvedFilePath),
                               fit: BoxFit.cover,
                               errorBuilder:
                                   (_, __, ___) => Container(